+++ src/test/scala/ai/starlake/utils/conversion/BigQueryUtilsSpec.scala
@@ -128,7 +128,8 @@ class BigQueryUtilsSpec extends TestHelper {
 
         val schemaHandler = new SchemaHandler(settings.storageHandler)
 
-        val schema = schemaHandler.domains
+        val schema = schemaHandler
+          .domains()
           .flatMap(_.tables)
           .find(_.name == "User")
           .map(_.bqSchema(schemaHandler))